clean up some types
