+++ public/app/features/query-library/api/types.ts
@@ -1,35 +1,3 @@
-import { DataQuery } from '@grafana/schema/dist/esm/index';
-
-export type DataQueryTarget = {
-  variables: object; // TODO: Detect variables in #86838
-  properties: DataQuery;
-};
-
-export type DataQuerySpec = {
-  title: string;
-  vars: object[]; // TODO: Detect variables in #86838
-  targets: DataQueryTarget[];
-};
-
-export type DataQueryFullSpec = {
-  apiVersion: string;
-  kind: string;
-  metadata: {
-    generateName: string;
-    name?: string;
-    creationTimestamp?: string;
-    annotations?: { [key: string]: string };
-  };
-  spec: DataQuerySpec;
-};
-
-export type DataQueryPartialSpec = Partial<DataQuerySpec>;
-
-export type DataQuerySpecResponse = {
-  apiVersion: string;
-  items: DataQueryFullSpec[];
-};
-
 // pkg/apis/iam/v0alpha1/types_display.go
 export type UserDataQueryResponse = {
   apiVersion: string;
